refactor(embeddings): close ownership boundaries
